Folder reorganization, search split into algo and struct
